Convert texture-coordinates to single precision
From reading around the internet I get the impression that double
precision floats are not only overkill (I know that much), it seems that
there is limited support for them, which could probably cause problems.
Since OpenGL is causing me enough trouble already (trying to use VBOs
right now) I'm willing to grasp at straws and eliminate any potential
source of problems.

+++ source/rust/client/ui/renderer.rs
@@ -139,10 +139,10 @@ fn draw_texture(Vec2(pos_x, pos_y): Vec2, texture: &Texture) {
 			0.0          , texture_height, 0.0,
 			0.0          , 0.0           , 0.0];
 		let texture_coordinates = [
-			1.0, 0.0,
-			1.0, 1.0,
-			0.0, 0.0,
-			0.0, 1.0];
+			1.0f32, 0.0f32,
+			1.0f32, 1.0f32,
+			0.0f32, 0.0f32,
+			0.0f32, 1.0f32];
 
 		gl::EnableClientState(gl::VERTEX_ARRAY);
 		gl::EnableClientState(gl::TEXTURE_COORD_ARRAY);
@@ -156,7 +156,7 @@ fn draw_texture(Vec2(pos_x, pos_y): Vec2, texture: &Texture) {
 				vertices.as_ptr() as *gl::types::GLvoid);
 			gl::TexCoordPointer(
 				2,
-				gl::DOUBLE,
+				gl::FLOAT,
 				0,
 				texture_coordinates.as_ptr() as *gl::types::GLvoid);
 		}
